Rename 'run.py' to 'solve.py'
The old 'run.py' will no longer be a top-level module when the
'adventkit' folder is made a package.  A new top-level 'run.py' file
will then be needed, which will be outside the 'adventkit' module and
which can import functions from the old 'run.py' file.  To avoid having
two files with the same name, the old 'run.py' is renamed.